requirement.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
django-phonenumber-field==6.0.0
phonenumbers==8.12.37
pkg-resources==0.0.0
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2